+++ npe2/_inspection/__init__.py
@@ -1,3 +1,4 @@
+from ._compile import compile
 from ._visitors import (
     NPE2PluginModuleVisitor,
     find_npe1_module_contributions,
@@ -8,4 +9,5 @@ __all__ = [
     "NPE2PluginModuleVisitor",
     "find_npe2_module_contributions",
     "find_npe1_module_contributions",
+    "compile",
 ]
